2006-02-09  Peter Dennis Bartok  <pbartok@novell.com>
	* ColorPalette.cs, EncoderParameters.cs, EncoderParameter.cs,
	  ImageAttributes.cs: 64bit cleanup and some simplifications

2006-02-09  Peter Dennis Bartok  <pbartok@novell.com>

	* PrintingServicesWin32.cs: Simplyfied the code, removing the need
	  for pointer calculations (part of 64bit cleanup)

2006-02-09  Peter Dennis Bartok  <pbartok@novell.com>

	* Image.cs, Bitmap.cs: Removed unneeded casts
	* Icon.cs: Fixed code to be 64bit safe


svn path=/trunk/mcs/; revision=56738

+++ mcs/class/System.Drawing/System.Drawing.Imaging/ImageAttributes.cs
@@ -266,7 +266,7 @@ namespace System.Drawing.Imaging
 					gdiclr.to = map[i].NewColor.ToArgb();
 				
 					Marshal.StructureToPtr (gdiclr, lpPointer, false);
-					lpPointer = (IntPtr) (lpPointer.ToInt32() + mapsize);						
+					lpPointer = (IntPtr) (lpPointer.ToInt64() + mapsize);						
 				}
 			
 				Status status = GDIPlus.GdipSetImageAttributesRemapTable (nativeImageAttr, 
@@ -383,7 +383,7 @@ namespace System.Drawing.Imaging
 					gdiclr.to = map[i].NewColor.ToArgb();
 				
 					Marshal.StructureToPtr (gdiclr, lpPointer, false);
-					lpPointer = (IntPtr) (lpPointer.ToInt32() + mapsize);						
+					lpPointer = (IntPtr) (lpPointer.ToInt64() + mapsize);						
 				}
 			
 				Status status = GDIPlus.GdipSetImageAttributesRemapTable (nativeImageAttr,
